+++ src/lib/keyring.js
@@ -169,7 +169,7 @@ class Keyring {
 
         if (_.get(this.decryptedVault, `importedWallets.${importedChain}`) !== undefined && this.decryptedVault.importedWallets[importedChain].data.some(element => element.address === address) == true) {
             isImportedAddress = true;
-        } else if (this.decryptedVault[chain] !== undefined && this.decryptedVault[chain].public.some(element => Web3.utils.toChecksumAddress(element.address) === Web3.utils.toChecksumAddress(address)) == true) {
+        } else if (this.decryptedVault[chain] !== undefined && this.decryptedVault[chain].public.some(element => element.address.toLowerCase() === address.toLowerCase()) == true) {
             isImportedAddress = false;
         } else {
             return { error: ERROR_MESSAGE.ADDRESS_NOT_PRESENT };
@@ -318,6 +318,12 @@ class Keyring {
                 return { response: signedMessage.message };
             }
 
+            if (Chains?.[this.chain]){
+                    const { signedMessage } = await this[this.chain].signMessage(data, address, privateKey);
+                    return { response: signedMessage };
+                
+            }
+
             return { error: ERROR_MESSAGE.UNSUPPORTED_NON_EVM_FUNCTIONALITY }
             
         }
@@ -355,7 +361,7 @@ class Keyring {
             return { error: ERROR_MESSAGE.INCORRECT_PIN };
         };
  
-        const { error, response } = await this.exportPrivateKey(rawTx.from.toLowerCase(), pin);
+        const { error, response } = await this.exportPrivateKey(rawTx.from, pin);
 
         if (error) {
             return { error };
@@ -383,7 +389,15 @@ class Keyring {
             return { response: signedTx };
         }
 
-        const { signedTransaction } = await this[this.chain].signTransaction(rawTx, privateKey);
+        if(isImported) {
+            const { signedTransaction } = await this[this.chain].signTransaction(rawTx, privateKey);
+            return { response: signedTransaction };
+        }
+        else{
+            const { signedTransaction } = await this[this.chain].signTransaction(rawTx);
+            return { response: signedTransaction };
+        }
+        
 
         return { response: signedTransaction };
     }
@@ -411,10 +425,11 @@ class Keyring {
 
         const activeChains = await this.getActiveChains();
 
-        const valuesToRemove = Object.keys(Chains.evmChains);
+        const evmChainList = Object.keys(Chains.evmChains);
 
-        const filteredChains = activeChains.response.filter(activeChains => !valuesToRemove.includes(activeChains.chain));
+        const filteredChains = activeChains.response.filter(activeChains => !evmChainList.includes(activeChains.chain));
 
+        //generate other chain's keyring instance and add accounts to it as per decrypted vault
         if (filteredChains.length > 0) {
             filteredChains.forEach(async (chainData) => {
                 const { response: mnemonic } = await this.exportMnemonic(pin);
@@ -436,6 +451,7 @@ class Keyring {
 
         const mnemonic = await helper.cryptography(decryptedVault.eth.private.encryptedMnemonic, pin, 'decryption');
 
+        // clearing vault state and adding new accounts as per decrypted vault
         const restoredVault = await this.keyringInstance.createNewVaultAndRestore(JSON.stringify(encryptionKey), mnemonic);
 
         const numberOfAcc = this.decryptedVault.eth.numberOfAccounts;
@@ -723,7 +739,7 @@ class Keyring {
 
         const activeChains = await this.getActiveChains();
 
-        const valuesToRemove = Object.keys(Chains.evmChains);
+        const evmChainList = Object.keys(Chains.evmChains);
 
         accounts.evm.generatedWallets = ({ ...decryptedVault.eth.public })
 
@@ -733,7 +749,7 @@ class Keyring {
             accounts.evm.importedWallets = ({ ...decryptedVault.importedWallets.evmChains.data });
         }
 
-        const filteredChains = activeChains.response.filter(activeChains => !valuesToRemove.includes(activeChains.chain));
+        const filteredChains = activeChains.response.filter(activeChains => !evmChainList.includes(activeChains.chain));
 
         let nonEvmAccs = [];
 
@@ -743,18 +759,12 @@ class Keyring {
 
             if (containsGenerated) {
                 nonEvmAccs = decryptedVault[chain].public.filter((address) => address.isDeleted === false);
-
-                let result = nonEvmAccs.map(a => { return { address: a.address, label: a.label }});
-
-                accounts[chain] = { generatedWallets: [ ...result ] };
+                accounts[chain] = { generatedWallets: { ...decryptedVault[chain].public } };
             }
             
             if (containsImported) {
                 nonEvmAccs = decryptedVault.importedWallets[chain].data.filter((address) => address.isDeleted === false);
-
-                let result = nonEvmAccs.map(a => { return { address: a.address, label: a.label }});
-
-                (accounts[chain] === undefined) ? accounts[chain] = { importedWallets: [ ...result ] } : accounts[chain].importedWallets = [ ...result ];
+               (accounts[chain] === undefined) ? accounts[chain] = { importedWallets: { ...decryptedVault.importedWallets[chain].data } } : accounts[chain].importedWallets = { ...decryptedVault.importedWallets[chain].data };
             }
         });
 
@@ -780,7 +790,7 @@ class Keyring {
             return { response: balance };
         }
 
-        const balance = await Chains[this.chain].getBalance(address, rpcUrl);
+        const balance = await Chains[this.chain].getBalance(address);
 
         return { response: balance }; 
     }
@@ -796,7 +806,7 @@ class Keyring {
             return { error: ERROR_MESSAGE.INCORRECT_PIN };
         };
 
-        const { error, response } = await this.exportPrivateKey(address.toLowerCase(), pin);
+        const { error, response } = await this.exportPrivateKey(address, pin);
 
         if (error) {
             return { error };
@@ -859,12 +869,17 @@ class Keyring {
                 return { error: ERROR_MESSAGE.ADDRESS_NOT_PRESENT };
             }
             if (typeof this.decryptedVault[chain].public[objIndex].label === 'string' || this.decryptedVault[chain].public[objIndex].label instanceof String){
-                const chains = Object.keys(Chains.evmChains);
-                let obj = chains.reduce(function(acc, curr) {
-                    acc[curr] = newLabel;
-                    return acc;
-                  }, {});
+                if(chain === 'eth') {
+                    const chains = Object.keys(Chains.evmChains);
+                    let obj = chains.reduce(function(acc, curr) {
+                        acc[curr] = newLabel;
+                        return acc;
+                    }, {});
                   this.decryptedVault[chain].public[objIndex].label = obj;
+                }
+                else {
+                    this.decryptedVault[chain].public[objIndex].label = newLabel;
+                }             
             }
             else{
             (chain === 'eth') ? this.decryptedVault[chain].public[objIndex].label[chainName] = newLabel : this.decryptedVault[chain].public[objIndex].label = newLabel;
@@ -881,19 +896,24 @@ class Keyring {
     }
 
     async resetAllImportedWallets(currentPin, newPin) {
-        const chain = (Chains.evmChains.hasOwnProperty(this.chain) || this.chain === 'ethereum') ? 'eth' : this.chain;
-        const importedChain = (chain === 'eth') ? 'evmChains' : chain;
 
-        if (_.get(this.decryptedVault, `importedWallets.${importedChain}`) === undefined) {
+        if (_.get(this.decryptedVault, `importedWallets`) === undefined) {
             return null;
         } 
 
-        let data = this.decryptedVault.importedWallets[importedChain].data
-        for(let i = 0; i < data.length; i++) {
-            let decryptedPrivKey = await helper.cryptography(data[i].privateKey, currentPin, 'decryption');
-            let encryptedPrivKey = await helper.cryptography(decryptedPrivKey, newPin, 'encryption');
-            this.decryptedVault.importedWallets[importedChain].data[i].privateKey = encryptedPrivKey
+        let importedChains  = Object.keys(this.decryptedVault.importedWallets)
+        
+        for (let importedChain of importedChains)
+        {
+            let data = this.decryptedVault.importedWallets[importedChain].data
+            for(let i = 0; i < data.length; i++) {
+                let decryptedPrivKey = await helper.cryptography(data[i].privateKey, currentPin, 'decryption');
+                let encryptedPrivKey = await helper.cryptography(decryptedPrivKey, newPin, 'encryption');
+                this.decryptedVault.importedWallets[importedChain].data[i].privateKey = encryptedPrivKey
+
+            }
         }
+        
 }