+++ CHANGELOG.md
@@ -559,3 +559,12 @@
 ### 1.30.08 (2023-10-09) Nightly version : only for experimental use
 
 * Updated pin while resetting imported wallets
+
+### 1.31.0 (2023-10-09) Nightly version : only for experimental use
+
+* Integrated upgraded bitcoin controller into vault
+* Enable backward compatibility in vault recovery for evm chains for undefined chain parameter in logs
+* Generate a default wallet for every chain when generating vault
+* Add fallback for recover vault - the default address to be recovered for every chain
+* Updated recover vault to create default wallet address for bitcoin for preexisting vaults
+* Updated test result for bitcoin